# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/properties/{id}/room-rack": {
      "get": {
        "operationId": "PropertiesController_roomRack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/v1/properties/{id}/room-types": {
      "get": {
        "operationId": "PropertiesController_roomTypes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/v1/rooms/{id}/status": {
      "patch": {
        "operationId": "PropertiesController_setRoomStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/api/v1/cashiering/shifts": {
      "get": {
        "operationId": "CashieringController_list",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      },
      "post": {
        "operationId": "CashieringController_open",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      }
    },
    "/api/v1/cashiering/shifts/{id}": {
      "get": {
        "operationId": "CashieringController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      }
    },
    "/api/v1/cashiering/shifts/{id}/movements": {
      "post": {
        "operationId": "CashieringController_movement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      }
    },
    "/api/v1/cashiering/shifts/{id}/close": {
      "post": {
        "operationId": "CashieringController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      }
    },
    "/api/v1/cashiering/shifts/{id}/approve": {
      "post": {
        "operationId": "CashieringController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cashiering"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lodgiva API",
    "description": "Hotel management platform API. All protected routes require a Bearer access token; the tenant is always derived from the token and never from the request body.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}